rca-insights-engine/requirements.txt
Flask==2.3.3
requests==2.31.0
prometheus-client==0.17.1
openai==1.95.0
python-dotenv==1.0.0